# nexus-trading-terminal/packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEXUS Trading Intelligence",
    "description": "\nNEXUS Trading Intelligence - local AI trading research terminal.\n\n* Every market object is tagged with its data mode (LIVE / DEMO). DEMO data is synthetic.\n* SIGNAL SCORE is a rules-based 0-100 quality score, never a probability.\n* All trading endpoints operate on a PAPER (simulated) broker. Live execution is disabled.\n* Errors are returned as `{\"error\": {\"code\", \"message\", \"request_id\"}}`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/news": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "News",
        "operationId": "news_api_news_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsOut"
                  },
                  "title": "Response News Api News Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/calendar": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Calendar",
        "operationId": "calendar_api_calendar_get",
        "parameters": [
          {
            "name": "hours_back",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 720,
              "minimum": 0,
              "default": 24,
              "title": "Hours Back"
            }
          },
          {
            "name": "hours_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 720,
              "minimum": 1,
              "default": 168,
              "title": "Hours Ahead"
            }
          },
          {
            "name": "impact",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(LOW|MEDIUM|HIGH)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Impact"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 8
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Calendar Api Calendar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/risk/{symbol}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Calendar Risk",
        "operationId": "calendar_risk_api_calendar_risk__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Calendar Risk Api Calendar Risk  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/heatmap": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Heatmap",
        "operationId": "heatmap_api_heatmap_get",
        "parameters": [
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(FOREX|CRYPTO|INDICES|COMMODITIES)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Class"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Heatmap Api Heatmap Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/correlations": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Correlations",
        "operationId": "correlations_api_correlations_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 250,
              "minimum": 20,
              "default": 60,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Correlations Api Correlations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "impact": {
            "type": "string",
            "title": "Impact"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "forecast": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast"
          },
          "actual": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "minutes_until": {
            "type": "number",
            "title": "Minutes Until"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event",
          "currency",
          "country",
          "time",
          "impact",
          "previous",
          "forecast",
          "actual",
          "source",
          "provider",
          "is_demo",
          "minutes_until"
        ],
        "title": "EventOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Original source URL; null for DEMO items (never fabricated)"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Countries"
          },
          "sentiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment"
          },
          "sentiment_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Source"
          },
          "sentiment_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Rationale"
          },
          "importance": {
            "type": "string",
            "title": "Importance"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "headline",
          "summary",
          "source",
          "url",
          "published_at",
          "symbols",
          "countries",
          "sentiment",
          "sentiment_source",
          "sentiment_rationale",
          "importance",
          "provider",
          "is_demo"
        ],
        "title": "NewsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}